4.21

+++ part4/controllers/blog.js
@@ -57,7 +57,18 @@ blogRouter.post('/', async (request, response) => {
 
 blogRouter.delete('/:id', async (request,response) => {
   console.log("delete router")
-  await Blog.findByIdAndDelete(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({error:'token invalid'})
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() === decodedToken.id.toString() ) {
+    await Blog.findByIdAndDelete(request.params.id)
+  } else {
+    return response.status(401).json({error:'Not the same user who created the blog'})
+  }
+
   response.status(204).end()
 })
